PA-4037 Create a SCA CLI package
