+++ .gitignore
@@ -230,3 +230,6 @@ fabric.properties
 
 # NuGet
 **/obj/Debug/**
+
+# PyPI
+.pypirc
