+++ setup.cfg
@@ -1,0 +1,49 @@
+[metadata]
+name = soos-sca
+author = SOOS
+author_email = dev@soos.io
+license_files = LICENSE
+description= Software security, simplified.
+long_description = file: README.md
+long_description_content_type = text/markdown
+url = https://github.com/soos-io/soos-ci-analysis-python
+project_urls =
+  Issues = https://github.com/soos-io/soos-ci-analysis-python/issues
+  Source = https://github.com/soos-io/soos-ci-analysis-python
+  About = https://soos.io
+  Register = https://app.soos.io/register
+# keywords = 
+
+# Classifiers help users find your project by categorizing it.
+# For a list of valid classifiers, see https://pypi.org/classifiers/
+classifiers =
+  Development Status :: 5 - Production/Stable
+  License :: OSI Approved :: MIT License
+  Intended Audience :: Developers
+  Topic :: Software Development :: Build Tools
+  Programming Language :: Python :: 3
+  Programming Language :: Python :: 3 :: Only
+  Programming Language :: Python :: 3.6
+  Programming Language :: Python :: 3.7
+  Programming Language :: Python :: 3.8
+  Programming Language :: Python :: 3.9
+  Programming Language :: Python :: 3.10
+
+[options]
+packages = find:
+package_dir =
+  =src
+python_requires = >=3.6, <4
+install_requires =
+  requests==2.26.0
+  certifi==2020.6.20
+  idna==2.10
+  urllib3==1.26.5
+  charset-normalizer==2.0.12
+
+[options.packages.find]
+where = src
+
+[options.entry_points]
+console_scripts =
+  soos-sca = cli.sca:entry_point
