+++ src/cli/__init__.py
@@ -1,0 +1,1 @@
+# intentionally left blank
