+++ src/cli/version.py
@@ -1,0 +1,4 @@
+__version__ = "1.6.0"
+
+def get_current_version():
+  return __version__
